Ganti tipe data ipk

+++ POSTTEST_1/soal2.cpp
@@ -5,7 +5,7 @@ struct Mahasiswa
 {
     string nama;
     string NIM;
-    double IPK;
+    float IPK;
 };
 
 int main(){
